Move token functions into endpoint module
