+++ quinn-proto/src/crypto/ring.rs
@@ -8,16 +8,8 @@ use ring::hkdf;
 use ring::hmac::{self, SigningKey};
 
 use crate::packet::{PacketNumber, LONG_HEADER_FORM};
-use crate::shared::{ConfigError, ConnectionId, ResetToken};
-use crate::{crypto, Side, RESET_TOKEN_SIZE};
-
-pub(crate) fn reset_token_for<H>(key: &H, id: &ConnectionId) -> ResetToken where H: crypto::HmacKey {
-    let signature = key.sign(id);
-    // TODO: Server ID??
-    let mut result = [0; RESET_TOKEN_SIZE];
-    result.copy_from_slice(&signature.as_ref()[..RESET_TOKEN_SIZE]);
-    result.into()
-}
+use crate::shared::{ConfigError, ConnectionId};
+use crate::{crypto, Side};
 
 /// Keys for encrypting and decrypting packet payloads
 pub struct Crypto {
@@ -265,88 +257,6 @@ const INITIAL_SALT: [u8; 20] = [
     0x48, 0x5e, 0x09, 0xa0,
 ];
 
-pub(crate) mod token {
-    use std::io;
-    use std::net::{IpAddr, SocketAddr};
-    use std::time::{Duration, SystemTime, UNIX_EPOCH};
-
-    use bytes::{Buf, BufMut};
-
-    use crate::coding::{BufExt, BufMutExt};
-    use crate::crypto::HmacKey;
-    use crate::shared::ConnectionId;
-    use crate::{MAX_CID_SIZE, MIN_CID_SIZE};
-
-    // TODO: Use AEAD to hide token details from clients for better stability guarantees:
-    // - ticket consists of (random, aead-encrypted-data)
-    // - AEAD encryption key is HKDF(master-key, random)
-    // - AEAD nonce is always set to 0
-    // in other words, for each ticket, use different key derived from random using HKDF
-
-    pub fn generate<K>(
-        key: &K,
-        address: &SocketAddr,
-        dst_cid: &ConnectionId,
-        issued: SystemTime,
-    ) -> Vec<u8>
-    where
-        K: HmacKey,
-    {
-        let mut buf = Vec::new();
-        buf.write(dst_cid.len() as u8);
-        buf.put_slice(dst_cid);
-        buf.write::<u64>(
-            issued
-                .duration_since(UNIX_EPOCH)
-                .map(|x| x.as_secs())
-                .unwrap_or(0),
-        );
-        let signature_pos = buf.len();
-        match address.ip() {
-            IpAddr::V4(x) => buf.put_slice(&x.octets()),
-            IpAddr::V6(x) => buf.put_slice(&x.octets()),
-        }
-        buf.write(address.port());
-        let signature = key.sign(&buf);
-        // No reason to actually encode the IP in the token, since we always have the remote addr for an incoming packet.
-        buf.truncate(signature_pos);
-        buf.extend_from_slice(signature.as_ref());
-        buf
-    }
-
-    pub fn check<K>(
-        key: &K,
-        address: &SocketAddr,
-        data: &[u8],
-    ) -> Option<(ConnectionId, SystemTime)>
-    where
-        K: HmacKey,
-    {
-        let mut reader = io::Cursor::new(data);
-        let dst_cid_len = reader.get::<u8>().ok()? as usize;
-        if dst_cid_len > reader.remaining()
-            || dst_cid_len != 0 && (dst_cid_len < MIN_CID_SIZE || dst_cid_len > MAX_CID_SIZE)
-        {
-            return None;
-        }
-        let dst_cid = ConnectionId::new(&data[1..=dst_cid_len]);
-        reader.advance(dst_cid_len);
-        let issued = UNIX_EPOCH + Duration::new(reader.get::<u64>().ok()?, 0);
-        let signature_start = reader.position() as usize;
-
-        let mut buf = Vec::new();
-        buf.put_slice(&data[0..signature_start]);
-        match address.ip() {
-            IpAddr::V4(x) => buf.put_slice(&x.octets()),
-            IpAddr::V6(x) => buf.put_slice(&x.octets()),
-        }
-        buf.write(address.port());
-
-        key.verify(&buf, &data[signature_start..]).ok()?;
-        Some((dst_cid, issued))
-    }
-}
-
 impl crypto::HmacKey for hmac::SigningKey {
     const KEY_LEN: usize = 64;
     type Signature = hmac::Signature;
@@ -373,7 +283,7 @@ mod test {
     use super::*;
     use crate::crypto::{HeaderKeys, Keys};
     use crate::MAX_CID_SIZE;
-    use rand::{self, RngCore};
+    use rand;
 
     #[test]
     fn handshake_crypto_roundtrip() {
@@ -449,25 +359,6 @@ mod test {
         let mut payload = BytesMut::from(payload);
         client.decrypt(0, &header, &mut payload).unwrap();
         assert_eq!(&payload, &plaintext[HEADER_LEN..]);
-    }
-
-    #[test]
-    fn token_sanity() {
-        use crate::crypto::HmacKey;
-        use crate::MAX_CID_SIZE;
-        use std::net::{Ipv6Addr, SocketAddr};
-        use std::time::{Duration, UNIX_EPOCH};
-
-        let mut key = [0; 64];
-        rand::thread_rng().fill_bytes(&mut key);
-        let key = <SigningKey as HmacKey>::new(&key).unwrap();
-        let addr = SocketAddr::new(Ipv6Addr::LOCALHOST.into(), 4433);
-        let dst_cid = ConnectionId::random(&mut rand::thread_rng(), MAX_CID_SIZE);
-        let issued = UNIX_EPOCH + Duration::new(42, 0); // Fractional seconds would be lost
-        let token = token::generate(&key, &addr, &dst_cid, issued);
-        let (dst_cid2, issued2) = token::check(&key, &addr, &token).expect("token didn't validate");
-        assert_eq!(dst_cid, dst_cid2);
-        assert_eq!(issued, issued2);
     }
 
     #[test]
